+++ packages/opentelemetry-resources/src/platform/browser/ServiceInstanceIdDetectorSyncA.ts
