+++ packages/opentelemetry-resources/src/platform/node/ServiceInstanceIdDetectorSyncA.ts
